refactor: aggregate cursusUsers with functions in services/
